3# description Fragment

+++ app/src/main/java/com/vodeg/airlines_app/data/model/Airline.kt
@@ -3,7 +3,6 @@ package com.vodeg.airlines_app.data.model
 
 import android.os.Parcelable
 import androidx.room.Entity
-import androidx.room.PrimaryKey
 import kotlinx.android.parcel.Parcelize
 import kotlinx.serialization.SerialName
 
@@ -19,7 +18,7 @@ data class Airline(
     @SerialName("head_quaters")
     val headQuaters: String?,
     @SerialName("id")
-    val id: Long?,
+    val id: Double?,
     @SerialName("logo")
     val logo: String?,
     @SerialName("name")
